backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
google-generativeai>=0.3.0
fpdf2>=2.7.4
numpy>=1.24.0
gemmi>=0.6.2
